+++ src/Contracts/CountryExceptionInterface.php
@@ -1,0 +1,22 @@
+<?php
+
+declare(strict_types=1);
+
+namespace MiBo\Countries\Contracts;
+
+use Throwable;
+
+/**
+ * Interface CountryExceptionInterface
+ *
+ * @package MiBo\Countries\Contracts
+ *
+ * @author Michal Boris <michal.boris27@gmail.com>
+ *
+ * @since 2.0.0
+ *
+ * @no-named-arguments Parameter names are not covered by the backward compatibility promise.
+ */
+interface CountryExceptionInterface extends Throwable
+{
+}