+++ src/Contracts/CountryProviderInterface.php
@@ -22,34 +22,46 @@ interface CountryProviderInterface
      *
      * @param non-empty-string $name
      *
-     * @return \MiBo\Countries\Contracts\CountryInterface|null
+     * @return \MiBo\Countries\Contracts\CountryInterface
+     *
+     * @throws \MiBo\Countries\Exceptions\CountryNotFoundException If no country with the given name was found.
+     * @throws \MiBo\Countries\Exceptions\InvalidSearchedValueException If provided value is invalid.
      */
-    public function getByName(string $name): ?CountryInterface;
+    public function getByName(string $name): CountryInterface;
 
     /**
      * Retrieves country by its alpha-2 code. (US, SK,...)
      *
      * @param non-empty-string $alpha2
      *
-     * @return \MiBo\Countries\Contracts\CountryInterface|null
+     * @return \MiBo\Countries\Contracts\CountryInterface
+     *
+     * @throws \MiBo\Countries\Exceptions\CountryNotFoundException If no country with the given alpha2 was found.
+     * @throws \MiBo\Countries\Exceptions\InvalidSearchedValueException If provided value is invalid.
      */
-    public function getByAlpha2(string $alpha2): ?CountryInterface;
+    public function getByAlpha2(string $alpha2): CountryInterface;
 
     /**
      * Retrieves country by its alpha-3 code. (USA, SVK,...)
      *
      * @param non-empty-string $alpha3
      *
-     * @return \MiBo\Countries\Contracts\CountryInterface|null
+     * @return \MiBo\Countries\Contracts\CountryInterface
+     *
+     * @throws \MiBo\Countries\Exceptions\CountryNotFoundException If no country with the given alpha3 was found.
+     * @throws \MiBo\Countries\Exceptions\InvalidSearchedValueException If provided value is invalid.
      */
-    public function getByAlpha3(string $alpha3): ?CountryInterface;
+    public function getByAlpha3(string $alpha3): CountryInterface;
 
     /**
      * Retrieves country by its numerical code.
      *
      * @param numeric-string $numericalCode
      *
-     * @return \MiBo\Countries\Contracts\CountryInterface|null
+     * @return \MiBo\Countries\Contracts\CountryInterface
+     *
+     * @throws \MiBo\Countries\Exceptions\CountryNotFoundException If no country with the given numeric was found.
+     * @throws \MiBo\Countries\Exceptions\InvalidSearchedValueException If provided value is invalid.
      */
-    public function getByNumericalCode(string $numericalCode): ?CountryInterface;
+    public function getByNumericalCode(string $numericalCode): CountryInterface;
 }