+++ CHANGELOG.md
@@ -5,3 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## 2.0
+
+### Added
+* CountryExceptionInterface - throwing on country provider errors such as missing country
+
+### Changed
+* CountryProvider - no longer nullable return type
+
+### Removed
+* ISO Currencies - Library is no longer native to this library
+
+### Fixed
+* ISO Country Provider - Implemented instance's cache